@@ -11,26 +11,32 @@
     "dev:rs": "nodemon src/rs-index.ts --ignore src/schema/"
   },
   "dependencies": {
-    "apollo-server-express": "^2.19.1",
+    "@apollo/server": "^4.5.0",
+    "@graphql-tools/schema": "^9.0.17",
     "axios": "^0.18.1",
+    "body-parser": "^1.20.2",
     "chokidar": "^3.5.0",
     "compression": "^1.7.4",
+    "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "graphql": "^14.2.0",
-    "graphql-tools": "^7.0.2",
-    "lodash": "^4.17.20"
+    "express": "^4.18.2",
+    "graphql": "^16.6.0",
+    "graphql-sse": "^2.1.0",
+    "graphql-subscriptions": "^2.0.0",
+    "graphql-upload": "15.0.2",
+    "graphql-ws": "^5.12.0",
+    "lodash": "^4.17.20",
+    "ws": "^8.13.0"
   },
   "devDependencies": {
     "@types/compression": "^1.7.0",
-    "@types/express-serve-static-core": "^4.17.7",
+    "@types/express": "^4.17.17",
     "@types/graphql": "^14.2.0",
     "@types/lodash": "^4.14.123",
-    "nodemon": "^2.0.7",
-    "ts-node": "^8.0.3",
-    "typescript": "3.4.1"
-  },
-  "resolutions": {
-    "@types/express-serve-static-core": "4.17.7"
+    "@types/node": "^18.15.0",
+    "@types/ws": "^8.5.4",
+    "nodemon": "^2.0.21",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   }
 }
