@@ -1,7 +1,6 @@
-import axios from 'axios';
-import { gql } from 'apollo-server-express';
+import axios from "axios";
 
-export const typeDef = gql`
+export const typeDef = `#graphql
   extend type Query {
     HNUser(username: String): HNUser
   }
@@ -14,12 +13,15 @@ export const typeDef = gql`
 
 export const resolvers = {
   Query: {
-    HNUser: (root: any, { username = 'imolorhe' }) =>
-      axios.get(`https://hacker-news.firebaseio.com/v0/user/${username}.json?print=pretty`)
-      .then(res => res.data)
-      .then(data => {
-        data.bio = data.about;
-        return data;
-      }),
-  }
+    HNUser: (root: any, { username = "imolorhe" }) =>
+      axios
+        .get(
+          `https://hacker-news.firebaseio.com/v0/user/${username}.json?print=pretty`
+        )
+        .then((res) => res.data)
+        .then((data) => {
+          data.bio = data.about;
+          return data;
+        }),
+  },
 };
